// pj-client/src/app/domain/models/Control.model.ts
export interface ControlModel{
    id_control?: number;
    id_responsable: number;
    id_expediente: number;
    observaciones: string | null;
    val_observaciones: boolean | null;
    val_datos: boolean | null;
    val_nitidez: boolean | null;
    val_pruebas_impresion: boolean | null;
    val_copia_fiel: boolean | null;
}